+++ R/search.R
@@ -164,10 +164,49 @@ search_L1 <- function(p_ref, refmodel, nterms_max, penalty, opt) {
     p_ref, nlist(x, weights = refmodel$wobs), refmodel$family,
     intercept = TRUE, ncol(x), penalty, opt
   )
-  solution_terms <- collapse_ranked_predictors(
+
+  solution_terms_orig <- collapse_ranked_predictors(
     path = colnames(x)[search_path$solution_terms], formula = refmodel$formula,
     data = fr
   )
+  solution_terms <- utils::head(solution_terms_orig, nterms_max)
+  # Place lower-order interaction terms before higher-order interaction terms,
+  # but otherwise preserve the ranking:
+  ia_orders <- sapply(gregexpr(":", solution_terms), function(greg_colon) {
+    sum(greg_colon != -1)
+  })
+  ia_order_max <- max(ia_orders)
+  for (ia_order in rev(seq_len(ia_order_max))) {
+    ias <- solution_terms[ia_orders == ia_order]
+    stopifnot(!any(duplicated(ias))) # safety measure for which.max()
+    for (ia in ias) {
+      ia_idx <- which.max(solution_terms == ia)
+      if (ia_idx > nterms_max) break
+      main_terms_ia <- strsplit(ia, ":")[[1]]
+      ias_lower_split <- utils::combn(main_terms_ia, m = ia_order,
+                                      simplify = FALSE)
+      ias_lower <- lapply(ias_lower_split, all_ia_perms, is_split = TRUE)
+      ias_lower <- unlist(ias_lower)
+      ias_lower <- intersect(ias_lower, solution_terms_orig)
+      prev_terms <- utils::head(solution_terms, ia_idx - 1L)
+      has_lower_after <- !all(ias_lower %in% prev_terms)
+      if (has_lower_after) {
+        if (getOption("projpred.warn_L1_interactions", TRUE)) {
+          warning("Interaction term `", ia, "` was selected before all ",
+                  "corresponding lower-order interaction terms have been ",
+                  "selected. This is a known deficiency of L1 search. Use ",
+                  "forward search to avoid this. Now ranking the lower-order ",
+                  "interaction terms before this interaction term.")
+        }
+        ias_lower <- setdiff(ias_lower, prev_terms)
+        ias_lower <- ias_lower[order(match(ias_lower, solution_terms_orig))]
+        new_head <- c(prev_terms, ias_lower, ia)
+        solution_terms <- c(new_head, setdiff(solution_terms, new_head))
+        solution_terms <- utils::head(solution_terms, nterms_max)
+      }
+    }
+  }
+
   outdmins <- lapply(0:length(solution_terms), function(nterms) {
     if (nterms == 0) {
       formula <- make_formula(c("1"))
@@ -210,21 +249,6 @@ search_L1 <- function(p_ref, refmodel, nterms_max, penalty, opt) {
     class(sub) <- "subfit"
     return(list(sub))
   })
-  solution_terms <- solution_terms[seq_len(nterms_max)]
-  outdmins <- outdmins[seq_len(nterms_max + 1)]
-
-  # Check for interaction terms being selected before all involved main effects
-  # have been selected (and throw a warning if that is the case):
-  ia_sel_bef_main <- sapply(grep(":", solution_terms), function(idx_ia) {
-    term_split <- strsplit(solution_terms[idx_ia], ":")[[1]]
-    !all(term_split %in% utils::head(solution_terms, idx_ia - 1L))
-  })
-  if (any(ia_sel_bef_main) &&
-      getOption("projpred.warn_L1_interactions", TRUE)) {
-    warning("An interaction has been selected before all involved main ",
-            "effects have been selected. This is a known deficiency of L1 ",
-            "search. Use forward search to avoid this.")
-  }
 
   return(nlist(solution_terms, outdmins))
 }